modify(entity, sql): 금액 관련 필드 long/BIGINT 타입으로 통일

+++ src/main/java/com/nhnacademy/back/pointpolicy/domain/entity/PointPolicy.java
@@ -24,7 +24,7 @@ public class PointPolicy {
 	private String pointPolicyName;
 
 	@Column(nullable = false)
-	private int pointPolicyFigure;
+	private long pointPolicyFigure;
 
 	@Column(nullable = false)
 	private LocalDateTime pointPolicyCreatedAt;
